requirements.txt
google-cloud-firestore==1.6.0
google-cloud-vision==0.41.0
instagram-scraper==1.7.1
requests==2.22.0
PyYAML==5.2
